+++ analysis/src/main/java/cz/vutbr/fit/group/verifit/oslc/analysis/properties/VeriFitAnalysisProperties.java
@@ -74,6 +74,11 @@ public class VeriFitAnalysisProperties
 		ADAPTER_PORT = VeriFitAnalysisProperties.getProperty("adapter_port");	
 		if (ADAPTER_PORT == null)
 			throw new IOException("adapter_port missing");
+		try { // check if parsable as int
+			Integer.parseInt(ADAPTER_PORT);
+		} catch (Exception e) {
+			throw new IOException("failed to parse adapter_port value - int expected");
+		}
 		
 		SPARQL_SERVER_QUERY_ENDPOINT = VeriFitAnalysisProperties.getProperty("sparql_query");	
 		if (SPARQL_SERVER_QUERY_ENDPOINT == null)
@@ -89,48 +94,51 @@ public class VeriFitAnalysisProperties
 		
 		String str_AUTHENTICAION_ENABLED;
 		str_AUTHENTICAION_ENABLED = VeriFitAnalysisProperties.getProperty("enable_authentication");	
-		if (str_AUTHENTICAION_ENABLED == null)
-			throw new IOException("enable_authentication missing");
-		try {
-			AUTHENTICATION_ENABLED = Boolean.parseBoolean(str_AUTHENTICAION_ENABLED);
-		} catch (Exception e) {
-			throw new IOException("failed to parse enable_authentication value - boolean expected");
+		if (str_AUTHENTICAION_ENABLED == null) {
+			AUTHENTICATION_ENABLED = false;	// assume false by default
+		} else {
+			try {
+				AUTHENTICATION_ENABLED = Boolean.parseBoolean(str_AUTHENTICAION_ENABLED);
+			} catch (Exception e) {
+				throw new IOException("failed to parse enable_authentication value - boolean expected");
+			}			
 		}
-		
-		AUTHENTICATION_USERNAME = VeriFitAnalysisProperties.getProperty("username");	
-		if (AUTHENTICATION_USERNAME == null)
-			throw new IOException("username missing");
-		
-		AUTHENTICATION_PASSWORD = VeriFitAnalysisProperties.getProperty("password");	
-		if (AUTHENTICATION_PASSWORD == null)
-			throw new IOException("password missing");
-		
+		if (AUTHENTICATION_ENABLED) { // load other relevant props only if needed
+			AUTHENTICATION_USERNAME = VeriFitAnalysisProperties.getProperty("username");	
+			if (AUTHENTICATION_USERNAME == null)
+				throw new IOException("username missing");
+			
+			AUTHENTICATION_PASSWORD = VeriFitAnalysisProperties.getProperty("password");	
+			if (AUTHENTICATION_PASSWORD == null)
+				throw new IOException("password missing");
+		}
 		
 		String str_KEEP_LAST_N_ENABLED;
 		str_KEEP_LAST_N_ENABLED = VeriFitAnalysisProperties.getProperty("keep_last_n_enabled");	
-		if (str_KEEP_LAST_N_ENABLED == null)
-			throw new IOException("keep_last_n_enabled missing");
-		try {
-			KEEP_LAST_N_ENABLED = Boolean.parseBoolean(str_KEEP_LAST_N_ENABLED);
-		} catch (Exception e) {
-			throw new IOException("failed to parse keep_last_n_enabled value - boolean expected");
-			
+		if (str_KEEP_LAST_N_ENABLED == null) {
+			KEEP_LAST_N_ENABLED = false;	// assume false by default
+		} else {
+			try {
+				KEEP_LAST_N_ENABLED = Boolean.parseBoolean(str_KEEP_LAST_N_ENABLED);
+			} catch (Exception e) {
+				throw new IOException("failed to parse keep_last_n_enabled value - boolean expected");
+			}
 		}
-		
-		String str_KEEP_LAST_N;
-		str_KEEP_LAST_N = VeriFitAnalysisProperties.getProperty("keep_last_n");	
-		if (str_KEEP_LAST_N == null)
-			throw new IOException("keep_last_n missing");
-		try {
-			KEEP_LAST_N = Long.parseLong(str_KEEP_LAST_N);
-			if (KEEP_LAST_N < 1)
-				throw new Exception();
-		} catch (Exception e) {
-			throw new IOException("failed to parse keep_last_n value - positive non-zero integer expected");
-			
+		if (KEEP_LAST_N_ENABLED) { // load other relevant props only if needed
+			String str_KEEP_LAST_N;
+			str_KEEP_LAST_N = VeriFitAnalysisProperties.getProperty("keep_last_n");	
+			if (str_KEEP_LAST_N == null)
+				throw new IOException("keep_last_n missing");
+			try {
+				KEEP_LAST_N = Long.parseLong(str_KEEP_LAST_N);
+				if (KEEP_LAST_N < 1)
+					throw new Exception();
+			} catch (Exception e) {
+				throw new IOException("failed to parse keep_last_n value - positive non-zero integer expected");
+				
+			}
 		}
 
-		
 		if (SystemUtils.IS_OS_LINUX) {
 			CONFIG_OS = ConfigOs.LINUX;
 			CONFIG_OS_STR = "linux";
@@ -150,6 +158,61 @@ public class VeriFitAnalysisProperties
 				throw new IOException("failed to parse config_win_shell value - options are \"windows_ps1\" or \"windows_bat\"");
 			}
 			CONFIG_OS_STR = str_CONFIG_OS;
+		}
+		
+		// AHT service registry configuration
+		String str_AHT_ENABLED;
+		str_AHT_ENABLED = VeriFitAnalysisProperties.getProperty("aht_enabled");	
+		if (str_AHT_ENABLED == null) {
+			AHT_ENABLED = false;	// assume false by default 
+		} else {
+			try {
+				AHT_ENABLED = Boolean.parseBoolean(str_AHT_ENABLED);
+			} catch (Exception e) {
+				throw new IOException("failed to parse aht_enabled value - boolean expected");
+			}
+		}
+		if (AHT_ENABLED) { // load other AHT props only if enabled
+			AHT_SERVICE_REGISTRY_HOST = VeriFitAnalysisProperties.getProperty("aht_service_registry_host");	
+			if (AHT_SERVICE_REGISTRY_HOST == null)
+				throw new IOException("aht_service_registry_host missing");
+			
+			AHT_SERVICE_REGISTRY_PORT = VeriFitAnalysisProperties.getProperty("aht_service_registry_port");	
+			if (AHT_SERVICE_REGISTRY_PORT == null)
+				throw new IOException("aht_service_registry_port missing");
+			try { // check if parsable as int
+				Integer.parseInt(AHT_SERVICE_REGISTRY_PORT);
+			} catch (Exception e) {
+				throw new IOException("failed to parse aht_service_registry_port value - int expected");
+			}
+			
+			AHT_SERVICE_NAME = VeriFitAnalysisProperties.getProperty("aht_service_name");	
+			if (AHT_SERVICE_NAME == null)
+				throw new IOException("aht_service_name missing");
+			
+			AHT_SYSTEM_NAME = VeriFitAnalysisProperties.getProperty("aht_system_name");	
+			if (AHT_SYSTEM_NAME == null)
+				throw new IOException("aht_system_name missing");
+			
+			AHT_CERTIFICATE = VeriFitAnalysisProperties.getProperty("aht_certificate");
+			if (AHT_CERTIFICATE == null)
+				throw new IOException("aht_certificate missing");
+			
+			AHT_CERTIFICATE_PASSWORD = VeriFitAnalysisProperties.getProperty("aht_certificate_password");
+			if (AHT_CERTIFICATE_PASSWORD == null)
+				throw new IOException("aht_certificate_password missing");
+		}
+		
+		String str_INPROGRESS_OUTPUTS_ENABLED;
+		str_INPROGRESS_OUTPUTS_ENABLED = VeriFitAnalysisProperties.getProperty("inprogress_outputs_enabled");	
+		if (str_INPROGRESS_OUTPUTS_ENABLED == null) {
+			INPROGRESS_OUTPUTS_ENABLED = false;	// assume false by default
+		} else {
+			try {
+				INPROGRESS_OUTPUTS_ENABLED = Boolean.parseBoolean(str_INPROGRESS_OUTPUTS_ENABLED);
+			} catch (Exception e) {
+				throw new IOException("failed to parse inprogress_outputs_enabled value - boolean expected");
+			}
 		}
 	}
 
@@ -188,6 +251,8 @@ public class VeriFitAnalysisProperties
 	public static final String PLUGIN_FILTER_CONF_PATH_BUILTIN = "./conf/BuiltInPluginFiltersConfiguration";
 	public static final String AUTOPLANS_DEF_PATH_CUSTOM = "./conf/CustomAnalysisTools";
 	public static final String PLUGIN_FILTER_CONF_PATH_CUSTOM = "./conf/CustomPluginFiltersConfiguration";
+	public static final String CERTIFICATES_PATH = "./conf/certificates";
+	public static final String UNIC_CONF_PATH = "./conf/unic";
 
 	public static String ADAPTER_HOST;
 	public static String ADAPTER_PORT;
@@ -207,13 +272,23 @@ public class VeriFitAnalysisProperties
 
 	public static ConfigOs CONFIG_OS;
 	public static String CONFIG_OS_STR;    
+
+	public static Boolean AHT_ENABLED;
+    public static String AHT_SERVICE_REGISTRY_HOST;
+    public static String AHT_SERVICE_REGISTRY_PORT;
+    public static String AHT_SERVICE_NAME;
+    public static String AHT_SYSTEM_NAME;
+    public static String AHT_CERTIFICATE;
+    public static String AHT_CERTIFICATE_PASSWORD;
+    
+    public static Boolean INPROGRESS_OUTPUTS_ENABLED;
     
     /*
      *  Internal constants
      */
 	public static final String AUTOMATION_PROVIDER_ID = "A0";
-	public static final String ADAPTER_CONTEXT = "analysis/";
-	public static String SERVER_URL = ADAPTER_HOST + ":" + ADAPTER_PORT + "/";
-    public static String PATH_AUTOMATION_SERVICE_PROVIDERS = SERVER_URL + ADAPTER_CONTEXT + "services/serviceProviders/";
-    public static String PATH_RESOURCE_SHAPES = SERVER_URL + ADAPTER_CONTEXT + "services/resourceShapes/";
+	public static final String ADAPTER_CONTEXT = "/analysis";
+	public static String SERVER_URL = ADAPTER_HOST + ":" + ADAPTER_PORT;
+    public static String PATH_AUTOMATION_SERVICE_PROVIDERS = SERVER_URL + ADAPTER_CONTEXT + "/services/serviceProviders";
+    public static String PATH_RESOURCE_SHAPES = SERVER_URL + ADAPTER_CONTEXT + "/services/resourceShapes";
 }